# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShareSlices API
  summary: ShareSlices account and Artifact HTTP API contract.
  description: >
    This OpenAPI document describes the current ShareSlices HTTP API contract
    for account entry and the version 0.0.1 Artifact upload, Preview, Publish,
    and Viewer flow, including current-browser-session sign out. Password
    recovery, email verification, additional account sign-in methods, link
    rotation, private sharing, and administration are outside this contract.

    The 429 responses are reserved for a future rate limiting change and are
    not implemented; contract tests exclude them until that change ships.
  version: 0.0.1
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Same-origin API server.
...
security: []
paths:
...
  /api/versions/{versionId}/content/:
    parameters:
      - $ref: "#/components/parameters/VersionId"
    get:
      tags:
        - preview
      operationId: previewVersionEntry
      summary: Preview a ready Version's root entry file.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Ready root entry streamed with caching disabled.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            text/html:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/VersionNotFound"
        "409":
          $ref: "#/components/responses/VersionNotReady"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/versions/{versionId}/content/{assetPath}:
    parameters:
      - $ref: "#/components/parameters/VersionId"
      - $ref: "#/components/parameters/AssetPath"
    get:
      tags:
        - preview
      operationId: previewVersionAsset
      summary: Preview one normalized committed manifest asset.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Manifest asset streamed with its recorded content type.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/AssetNotFound"
        "409":
          $ref: "#/components/responses/VersionNotReady"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /gallery-content/public/{playerAuthorization}/:
    parameters: [{$ref: "#/components/parameters/GalleryPlayerAuthorization"}]
    get:
      tags: [gallery-content]
      operationId: getGalleryContentEntry
      summary: Stream the credential-bound committed entry from the content-only runtime.
      responses:
        "200": {$ref: "#/components/responses/GalleryContentRead"}
        "404": {$ref: "#/components/responses/PublicGalleryNotFound"}
        "503": {$ref: "#/components/responses/PublicGalleryUnavailable"}
...
  /gallery-content/review/{reviewAuthorization}/:
    parameters: [{$ref: "#/components/parameters/GalleryReviewAuthorization"}]
    get:
      tags: [gallery-content]
      operationId: getGalleryReviewContentEntry
      summary: Stream a case-bound candidate entry using a distinct review credential.
      responses:
        "200": {$ref: "#/components/responses/GalleryContentRead"}
        "404": {$ref: "#/components/responses/PublicGalleryNotFound"}
        "503": {$ref: "#/components/responses/GalleryGovernanceUnavailable"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shareslices_session
      description: Browser session cookie established by successful email/password login.
    sessionBearer:
      type: http
      scheme: bearer
      description: Opaque independently revocable CLI Session credential.
  parameters:
...
    VersionId:
      name: versionId
      in: path
      required: true
      description: Stable immutable Version ID.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    AssetPath:
      name: assetPath
      in: path
      required: true
      description: >
        Greedy relative Artifact path. The server decodes, normalizes, and
        resolves it only against the committed manifest.
      schema:
        type: string
        minLength: 1
...
    GalleryPlayerAuthorization:
      name: playerAuthorization
      in: path
      required: true
      description: Short-lived opaque public credential; redact the value from every log.
      schema: {type: string, minLength: 32, maxLength: 512}
    GalleryReviewAuthorization:
      name: reviewAuthorization
      in: path
      required: true
      description: Short-lived opaque Administrator-review credential; never valid on public routes.
      schema: {type: string, minLength: 32, maxLength: 512}
...
  headers:
    RequestId:
      description: Stable request identifier for logs and support.
      schema:
        type: string
        minLength: 1
...
    NoStore:
      description: Prevents storage of sensitive or state-dependent responses.
      required: true
      schema:
        type: string
        enum:
          - no-store
    NoReferrer:
      description: Prevents credential-bearing content URLs from becoming referrer data.
      required: true
      schema: {type: string, enum: [no-referrer]}
...
  responses:
...
    Unauthenticated:
      description: Signed-in state is missing or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthenticated:
              value:
                error:
                  code: unauthenticated
                  message: Sign in to continue.
                  requestId: req_000000000000000000000000
...
    VersionNotFound:
      description: Version is unknown or is not owned by the current user.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: version_not_found
                  message: Version not found.
                  requestId: req_000000000000000000000000
    AssetNotFound:
      description: Normalized asset is absent from the committed Version manifest.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: asset_not_found
                  message: Asset not found.
                  requestId: req_000000000000000000000000
...
    VersionNotReady:
      description: Version exists but has not reached the ready state.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notReady:
              value:
                error:
                  code: version_not_ready
                  message: Version is not ready.
                  requestId: req_000000000000000000000000
...
    InternalError:
      description: The API encountered an unexpected failure.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              value:
                error:
                  code: internal_error
                  message: Internal server error.
                  requestId: req_000000000000000000000000
...
    GalleryContentRead:
      description: Bound entry or manifest asset with content-runtime security policy.
      headers:
        Cache-Control: {$ref: "#/components/headers/NoStore"}
        Referrer-Policy: {$ref: "#/components/headers/NoReferrer"}
        Content-Security-Policy: {schema: {type: string}}
        Permissions-Policy: {schema: {type: string}}
        X-Content-Type-Options: {schema: {type: string, enum: [nosniff]}}
      content: {application/octet-stream: {schema: {type: string, format: binary}}}
...
    PublicGalleryUnavailable:
      description: Pre-lookup response while Gallery is disabled or deployment-ineligible.
      headers: {X-Request-Id: {$ref: "#/components/headers/RequestId"}, Cache-Control: {$ref: "#/components/headers/NoStore"}, Referrer-Policy: {$ref: "#/components/headers/NoReferrer"}}
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}, examples: {unavailable: {value: {error: {code: gallery_unavailable, message: Gallery is unavailable., requestId: req_000000000000000000000000}}}}}}
    GalleryGovernanceUnavailable:
      description: A trusted governance dependency required by this operation is unavailable.
      headers: {X-Request-Id: {$ref: "#/components/headers/RequestId"}, Cache-Control: {$ref: "#/components/headers/NoStore"}}
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
    PublicGalleryNotFound:
      description: Generic response for unknown, Pending, Removed, Restricted, taken-down, or otherwise ineligible listing state.
      headers: {X-Request-Id: {$ref: "#/components/headers/RequestId"}, Cache-Control: {$ref: "#/components/headers/NoStore"}, Referrer-Policy: {$ref: "#/components/headers/NoReferrer"}}
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorResponse"}}}
...
  schemas:
...
    ValidationDetails:
      type: object
      additionalProperties: false
      properties:
        path:
          type: string
        paths:
          type: array
          items:
            type: string
        candidates:
          type: array
          items:
            type: string
        extension:
          type: string
        validationKind:
          type: string
        actualBytes:
          oneOf:
            - type: integer
              format: int64
              minimum: 0
              maximum: 9007199254740991
            - type: string
              pattern: "^(0|[1-9][0-9]*)$"
        limitBytes:
          oneOf:
            - type: integer
              format: int64
              minimum: 0
              maximum: 9007199254740991
            - type: string
              pattern: "^(0|[1-9][0-9]*)$"
        actualCount:
          oneOf:
            - type: integer
              format: int64
              minimum: 0
              maximum: 9007199254740991
            - type: string
              pattern: "^(0|[1-9][0-9]*)$"
        limitCount:
          oneOf:
            - type: integer
              format: int64
              minimum: 0
              maximum: 9007199254740991
            - type: string
              pattern: "^(0|[1-9][0-9]*)$"
        ignoredCount:
          oneOf:
            - type: integer
              format: int64
              minimum: 0
              maximum: 9007199254740991
            - type: string
              pattern: "^(0|[1-9][0-9]*)$"
        directory:
          type: string
        entryFile:
          type: string
    CliCompatibilityDetails:
      type: object
      additionalProperties: false
      required:
        - currentVersion
        - minimumVersion
        - operatingSystem
        - supportedOperatingSystems
      properties:
        currentVersion:
          type: string
        minimumVersion:
          type: string
          pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+$"
        operatingSystem:
          type: string
        supportedOperatingSystems:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [linux, macos, windows]
    RequestFieldDetails:
      type: object
      additionalProperties: false
      properties:
        expected:
          type: string
        received:
          type: string
    SizeLimitDetails:
      type: object
      additionalProperties: false
      required:
        - limitBytes
      properties:
        actualBytes:
          type: integer
          format: int64
          minimum: 0
        limitBytes:
          type: integer
          format: int64
          minimum: 1
    ConflictDetails:
      type: object
      additionalProperties: false
      properties:
        resourceId:
          type: string
        state:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
          description: Stable request identifier matching the X-Request-Id response header.
        action:
          type: string
          minLength: 1
        details:
          oneOf:
            - $ref: "#/components/schemas/ValidationDetails"
            - $ref: "#/components/schemas/CliCompatibilityDetails"
            - $ref: "#/components/schemas/RequestFieldDetails"
            - $ref: "#/components/schemas/SizeLimitDetails"
            - $ref: "#/components/schemas/ConflictDetails"
        fields:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      additionalProperties: false
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          minLength: 1
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1